requirements.txt
# Core AWS SDK
boto3==1.34.131
botocore==1.34.131

# AI/ML Models
transformers==4.46.3
torch==2.2.0
torchvision==0.17.0
accelerate==0.27.0
sentencepiece>=0.2.0

# LangChain
langchain==0.3.0
langchain-aws==0.2.0
langchain-core==0.3.0

# Vector Search
opensearch-py==2.4.0
requests-aws4auth==1.2.3

# Document Processing
pdf2image==1.16.3
pypdf==5.7.0
pillow==10.1.0
python-magic==0.4.27

# Date/Time
python-dateutil==2.8.2
holidays==0.39

# Data Processing
pandas==2.2.0
numpy==1.26.4

# API Framework
fastapi==0.110.0
mangum==0.17.0
pydantic==2.11.5

# Utilities
requests==2.32.3
python-dotenv==1.0.1
packaging==24.2
click==8.1.7

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
moto==4.2.0

# Monitoring
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0